analyst_app/requirements.txt
groq
sentence-transformers
faiss-cpu
numpy